Refactoring to make role more modular
Refractor the tasks and split them in multiple files / folders. I think,
this makes the role easier to read and extend. This also makes it easier,
to add a more granular controll for ufw. The new variables
`nginx_ufw_skip` and `nginx_ufw_install` allow you to controll, wether
if ufw is executed at all and if the corresponding packages will be
installed.